==> 6. Filtering 1 - bi-gram selection.ipynb <==
# 6. Filtering 1 - bi-gram selection

#### This script loads the bi-gram datasets and selects the most relevant ones, so as to not have too many sprseness in the dat

### Imports

import pandas as pd
import numpy as np

### Definitions

pos_all_ngram_filename = 'data/POS-ngram-all.csv'
bleached_all_ngram_filename = 'data/Bleached-ngram-all.csv'

pos_selected_ngram_filename = 'data/POS-ngram-selected.csv'
bleached_selected_ngram_filename = 'data/Bleached-ngram-selected.csv'

### Load POS bigrams

df = pd.read_csv(pos_all_ngram_filename)

df.head()

### Get information on what is considered as 'low frequency'

total_count = np.asarray(df['FrequencyMale'] + df['FrequencyFemale'])

for i in [1, 10, 20, 50, 100, 1000]:
    print('Counts more than {0}: {1}'.format(i, np.sum(total_count > i)))

We will consider only those more than 100

### Remove columns with low frequency

df = df[total_count > 100]

total_count = total_count[total_count > 100]

### Calculate the entropy column

male_prob, female_prob = df['FrequencyMale']/total_count, df['FrequencyFemale']/total_count

male_prob.shape
df['Entropy'] = - ( male_prob * np.log2(male_prob) + female_prob * np.log2(female_prob) )

df.head()

### Take top 100 with least entropy (ie which discriminate the most)

df = df.sort_values('Entropy').reset_index(drop = True)

df = df.iloc[:100]

df.head()

### Save to new CSV

df[['Word1', 'Word2']].to_csv(pos_selected_ngram_filename, index=False)

## Now we do the same but with the bleached n-grams instead

### Load Bleached bigrams

df = pd.read_csv(bleached_all_ngram_filename)

df.head()

### Get information on what is considered as 'low frequency'

total_count = np.asarray(df['FrequencyMale'] + df['FrequencyFemale'])

for i in [1, 10, 20, 50, 100, 1000]:
    print('Counts more than {0}: {1}'.format(i, np.sum(total_count > i)))

We will consider only those more than 100

### Remove columns with low frequency

df = df[total_count > 100]

total_count = total_count[total_count > 100]

### Calculate the entropy column

male_prob, female_prob = df['FrequencyMale']/total_count, df['FrequencyFemale']/total_count

male_prob.shape
df['Entropy'] = - ( male_prob * np.log2(male_prob) + female_prob * np.log2(female_prob) )

df.head()

### Take top 100 with least entropy (ie which discriminate the most)

df = df.sort_values('Entropy').reset_index(drop = True)

df = df.iloc[:100]

df.head()

### Save to new CSV

df[['Word1', 'Word2']].to_csv(pos_selected_ngram_filename, index=False)
