==> 7. Feature Extraction 4 - Function Words.ipynb <==
# 7. Feature Extraction 4 - Function Words

#### Extracting function words from the train set and storing the frequencies in a CSV file to later select the most suitable ones

### Imports

import nltk
import pandas as pd

### Definitions

post_list_dictionary = 'PostsList.csv'

function_word_list_filename = 'FunctionWords-all.txt'

function_word_extracted = 'FunctionWords-extracted.csv'

### Read the function word list into a dictionary

word_dict = {}

with open(function_word_list_filename, 'r') as f:
    words = f.readlines()

### Open the CSV file, go trought the training data and put the frequencies into the dictionary

number_of_posts=0
chunksize=10
for chunk in pd.read_csv(pos_filename, usecols=['PostID'], chunksize=chunksize):
    number_of_posts = np.max(chunk.PostID)

number_of_chunks = math.ceil(number_of_posts / chunksize)

for word in lines:
    word_dict[word] = (0, 0) # (Male Frequency, Female Frequency)
    
# open CSV file in append mode
with open(pos_ngram_filename, 'a') as f:
    for i,chunk in enumerate(pd.read_csv(bleached_filename, chunksize=chunksize)):
        chunk = chunk.dropna()
        
        if i%30 == 0:
            clear_output()
            print('{0}/{1}'.format(i+1, number_of_chunks), end = '', flush=True)
        
        for row in chunk.iterrows():
            if row[1][4] == 'train':
                post = row[1][3]
                words = nltk.word_tokenize(post)
                if len(words) > 0:
                    inverse_number_of_words = 1/len(words)
                    if row[1][2] == 'male':
                        for word in words:
                            if word in word_dict:
                                word_dict[word] = (
                                    word_dict[word][0] + inverse_number_of_words,
                                    word_dict[word][1]
                                )
                    else:
                        for word in words:
                            if word in word_dict:
                                word_dict[word] = (
                                    word_dict[word][0],
                                    word_dict[word][1] + inverse_number_of_words
                                )
clear_output()
print('DONE - {0} files'.format(number_of_posts))

df = pd.DataFrame(columns=['Word', 'FrequencyMale', 'FrequencyFemale'])

for Word1, (FrequencyMale, FrequencyFemale) in ngrams_dict.items():
    df = df.append({
        'Word': Word,
        'FrequencyMale': FrequencyMale,
        'FrequencyFemale': FrequencyFemale
    }, ignore_index = True)

df.to_csv(function_word_extracted, index=False)
